adding ware house product quntity form

+++ Warehouse Managment/Warehousess/ListWarehouses.cs
@@ -18,7 +18,7 @@ namespace Warehouse_Managment.Warehousess
         public ListWarehouses()
         {
             InitializeComponent();
-            dataGridView1.DataSource = db.Warehouses.ToList();
+            dataGridView1.DataSource = db.showstoresss().ToList();
             
 
         }
@@ -38,23 +38,41 @@ namespace Warehouse_Managment.Warehousess
 
         private void Button2_Click(object sender, EventArgs e)
         {
+            if (textBox3.Text != string.Empty)
+            {
 
-            var warhouses = db.Warehouses.SingleOrDefault(X => X.Warhouse_Id == id);
-            warhouses.Warhouse_Name = textBox3.Text;
-            warhouses.Warhouse_Address = textBox4.Text;
-            warhouses.Warhouse_Managers = textBox5.Text;
-            db.SaveChanges();
-            dataGridView1.DataSource = db.Warehouses.ToList();
-            MessageBox.Show("store updated");
+                var warhouses = db.Warehouses.SingleOrDefault(X => X.Warhouse_Id == id);
+                warhouses.Warhouse_Name = textBox3.Text;
+                warhouses.Warhouse_Address = textBox4.Text;
+                warhouses.Warhouse_Managers = textBox5.Text;
+                db.SaveChanges();
+                dataGridView1.DataSource = db.showstoresss().ToList();
+                MessageBox.Show("store updated");
+            }
+            else
+            {
+
+                MessageBox.Show("select store from table");
+            }
         }
 
         private void Button3_Click(object sender, EventArgs e)
         {
-            var result = db.Warehouses.Find(id);
-            db.Warehouses.Remove(result);
-            db.SaveChanges();
-            dataGridView1.DataSource = db.Warehouses.ToList();
-            MessageBox.Show("deleted");
+            if (textBox3.Text != string.Empty)
+            {
+                var result = db.Warehouses.Find(id);
+                db.Warehouses.Remove(result);
+                db.SaveChanges();
+                dataGridView1.DataSource = db.showstoresss().ToList();
+                MessageBox.Show("deleted");
+
+            }
+            else
+            {
+                MessageBox.Show("select store from table");
+            }
+
+
 
         }
